refactor: 향후 스키마 파일 분리를 위한 디렉토리 구조 재편
