+++ src/db/client.ts
@@ -1,7 +1,7 @@
 import { Pool } from 'pg';
 
-import * as schema from './schema';
 import { drizzle } from 'drizzle-orm/node-postgres';
+import * as schema from '../schema';
 
 export function createPool(databaseUrl: string): Pool {
   return new Pool({
