+++ src/db/index.ts
@@ -1,4 +1,4 @@
 export * from './client';
 export * from './db.constants';
 export * from './db.module';
-export * from './schema';
+export * from '../schema';
